app/requirements.txt
fastapi
uvicorn
httpx
beautifulsoup4
playwright
openai
python-multipart
pydantic
sqlalchemy
dotenv
google-search-results
google-generativeai